Fix examples for compilation problems
